backend/requirements.txt
asgiref==3.3.4
click==8.0.1
fastapi==0.65.1
h11==0.12.0
mongoengine==0.23.1
pydantic==1.8.2
PyJWT==2.1.0
pymongo==3.11.4
python-dotenv==0.17.1
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
yapf==0.31.0
